Agrego componente perfil-usuario, link en nav y llevo la sección de descarga de aplicación al final de home
