+++ frontend/src/app/pages/dashboard/dashboard.component.html
@@ -1,11 +1,11 @@
 <div class="d-flex h-100">
  
   <aside class="d-flex flex-column p-3 border-end" style="min-width: 200px;">
-    <button class="btn mb-2 btn-outline-primary" 
+    <!-- <button class="btn mb-2 btn-outline-primary" 
             [ngClass]="{ 'active': showPerfil }"
             (click)="showPerfilComponent()">
       👤 Perfil
-    </button>
+    </button> -->
     <button class="btn btn-outline-secondary"
             [ngClass]="{ 'active': showHistorial }"
             (click)="showHistorialComponent()">
