--- frontend/src/app/pages/dashboard/perfil-usuario/perfil-usuario.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PerfilUsuarioComponent } from './perfil-usuario.component';
-
-describe('PerfilUsuarioComponent', () => {
-  let component: PerfilUsuarioComponent;
-  let fixture: ComponentFixture<PerfilUsuarioComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [PerfilUsuarioComponent]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(PerfilUsuarioComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
