+++ frontend/src/app/pages/home/home.component.html
@@ -37,39 +37,6 @@
   </div>
 
 
-  <!-- Descargar app y github -->
-  <section class="app-download-section">
-    <div class="container-fluid">
-      <div class="row">
-        <!-- Texto a la izquierda -->
-        <div class="col-lg-6 col-md-6">
-          <div class="app-text-box">
-            <h2 class="section-title">¡Tu próxima hamburguesa,<br> a un solo clic!</h2>
-            <p class="section-subtitle">
-              DESCARGÁ NUESTRA APP Y DISFRUTÁ DE PROMOCIONES EXCLUSIVAS.<br>
-              ¡HACÉ TU PEDIDO EN SEGUNDOS ESTÉS DONDE ESTÉS!
-            </p>
-            <div class="button-group">
-              <a class="btn btn-success download-button" href="https://play.google.com/store/apps?hl=es_AR">
-                <i class="bi bi-google-play"></i> DESCARGAR EN PLAY STORE
-              </a>
-              <a class="btn btn-warning github-button" href="https://github.com/G10-ISPC/Backend">
-                <i class="bi bi-github"></i> VER EN GITHUB
-              </a>
-            </div>
-          </div>
-        </div>
-  
-        <!-- Columna vacía -->
-        <div class="col-lg-6 d-none d-lg-block">
-          <!--para dejar espacio a la derecha -->
-        </div>
-      </div>
-    </div>
-  </section>
-
-
-
 <!-- Download App Section -->
 <!-- <section class="download-app-section py-5">
   <div class="container">
@@ -122,9 +89,43 @@
           <p>
             Nos especializamos en la creación de productos gourmet con ingredientes frescos y naturales. En RICCO BURGERS, cuidamos los detalles para ofrecerte una experiencia única.
           </p>
-          <a routerLink="/aboutUs" class="btn btn-outline-light btn-hover-yellow">Conócenos más</a>
+          <a routerLink="/nosotros" class="btn btn-outline-light btn-hover-yellow">Conócenos más</a>
         </div>
       </div>
     </div>
   </aside>  
+
+
+
+    <!-- Descargar app y github -->
+  <section class="app-download-section">
+    <div class="container-fluid">
+      <div class="row">
+        <!-- Texto a la izquierda -->
+        <div class="col-lg-6 col-md-6">
+          <div class="app-text-box">
+            <h2 class="section-title">¡Tu próxima hamburguesa,<br> a un solo clic!</h2>
+            <p class="section-subtitle">
+              DESCARGÁ NUESTRA APP Y DISFRUTÁ <br>DE PROMOCIONES EXCLUSIVAS.<br>
+              ¡HACÉ TU PEDIDO EN SEGUNDOS <br>ESTÉS DONDE ESTÉS!
+            </p>
+            <div class="button-group">
+              <a class="btn btn-success download-button" href="https://play.google.com/store/apps?hl=es_AR">
+                <i class="bi bi-google-play"></i> DESCARGAR EN PLAY STORE
+              </a>
+              <a class="btn btn-warning github-button" href="https://github.com/G10-ISPC/Backend">
+                <i class="bi bi-github"></i> VER EN GITHUB
+              </a>
+            </div>
+          </div>
+        </div>
+  
+        <!-- Columna vacía -->
+        <div class="col-lg-6 d-none d-lg-block">
+          <!--para dejar espacio a la derecha -->
+        </div>
+      </div>
+    </div>
+  </section>
+
 </main>
