+++ frontend/src/app/pages/perfil-usuario/perfil-usuario.component.html
@@ -1,0 +1,57 @@
+ <div class="profile-wrapper">
+  <div class="col-md-4">
+    <div class="card profile-card shadow">
+      <div class="card-header">
+        <h5 class="mb-0">Editar Perfil</h5>
+      </div>
+      <div class="card-body">
+        <form [formGroup]="profileForm" (ngSubmit)="updateProfile()">
+          <div class="form-group mb-3">
+            <label for="first_name">Nombre</label>
+            <input id="first_name" type="text" class="form-control" formControlName="first_name" />
+          </div>
+
+          <div class="form-group mb-3">
+            <label for="last_name">Apellido</label>
+            <input id="last_name" type="text" class="form-control" formControlName="last_name" />
+          </div>
+
+          <div class="form-group mb-3">
+            <label for="telefono">Telefono</label>
+            <input id="telefono" type="telefono" class="form-control" formControlName="telefono" />
+          </div>
+          
+          <div class="form-group mb-3">
+            <label for="password">Contraseña</label>
+            <input id="password" type="password" class="form-control" formControlName="password" />
+          </div>
+          
+          <div class="form-group mb-3">
+            <label for="password2">Repetir Contraseña</label>
+            <input id="password2" type="password" class="form-control" formControlName="password2" />
+            <div *ngIf="profileForm.value.password && profileForm.value.password2 && profileForm.value.password !== profileForm.value.password2" class="text-danger mt-1">
+              Las contraseñas no coinciden.
+            </div>
+          </div>
+          
+          <div class="form-group mb-3">
+            <label for="email">Correo Electrónico</label>
+            <input id="email" type="email" class="form-control" formControlName="email" />
+          </div>
+
+          <button type="submit" class="btn btn-warning w-100" [disabled]="profileForm.invalid">
+            Guardar Cambios
+          </button>
+        </form>
+      </div>
+    </div>
+
+    <div class="text-center mt-3">
+      <p class="text-muted">
+        ¿Querés eliminar tu cuenta?
+        <a href="#" class="text-danger" (click)="deleteAccount()">Hacé clic acá</a>
+      </p>
+    </div>
+  </div>
+</div>
+
