+++ frontend/src/styles.css
@@ -14,13 +14,13 @@ body {
 
 
   .app-download-section {
-    background-image: url('/assets/img/riccomobile.png');
+    background-image: url('/assets/img/riccomobile1.png');
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center bottom bottom;
     background-color: #000;
     padding-top: 70px;
-    min-height: 85vh;
+    min-height: 80vh;
     width: 100%;
   }
 
@@ -29,15 +29,17 @@ body {
     padding: 30px 40px; 
     border-radius: 12px;
     color: white;
+    margin-top: -30px;
     margin-left: 40px;
     max-width: 450px; 
   }
   
   
   .section-title {
-    font-size: 1.6rem; 
+    font-size: 1.8 rem; 
     font-weight: 800;
     text-transform: uppercase;
+    margin-left: 15px;
     margin-bottom: 15px;
     color: #fff;
   }
@@ -49,7 +51,7 @@ body {
     margin-bottom: 20px;
     color: #ddd;
     text-align: center; 
-    margin-left: 20px;
+    margin-left: 15px;
   }
   
   .button-group {
